move otel setup
