+++ cmd/tracing.go
@@ -1,4 +1,4 @@
-package server
+package cmd
 
 import (
 	"context"
@@ -22,9 +22,6 @@ import (
 func setupOTelSDK(ctx context.Context) (shutdown func(context.Context) error, err error) {
 	var shutdownFuncs []func(context.Context) error
 
-	// shutdown calls cleanup functions registered via shutdownFuncs.
-	// The errors from the calls are joined.
-	// Each registered cleanup will be invoked once.
 	shutdown = func(ctx context.Context) error {
 		var err error
 		for _, fn := range shutdownFuncs {
@@ -34,20 +31,16 @@ func setupOTelSDK(ctx context.Context) (shutdown func(context.Context) error, er
 		return err
 	}
 
-	// handleErr calls shutdown for cleanup and makes sure that all errors are returned.
 	handleErr := func(inErr error) {
 		err = errors.Join(inErr, shutdown(ctx))
 	}
 
-	// Set up resource.
 	res, err := resource.New(ctx,
 		resource.WithAttributes(
-			// The service name used to display traces in backends
-			semconv.ServiceNameKey.String(name),
+			semconv.ServiceNameKey.String("k8s-dqlite"),
 		),
 	)
 
-	// Set up trace provider.
 	tracerProvider, err := newTraceProvider(ctx, res)
 	if err != nil {
 		handleErr(err)
@@ -56,7 +49,6 @@ func setupOTelSDK(ctx context.Context) (shutdown func(context.Context) error, er
 	shutdownFuncs = append(shutdownFuncs, tracerProvider.Shutdown)
 	otel.SetTracerProvider(tracerProvider)
 
-	// Set up meter provider.
 	meterProvider, err := newMeterProvider(res)
 	if err != nil {
 		handleErr(err)
@@ -68,8 +60,6 @@ func setupOTelSDK(ctx context.Context) (shutdown func(context.Context) error, er
 	return
 }
 
-// Initialize a gRPC connection to be used by both the tracer and meter
-// providers.
 func initConn() (*grpc.ClientConn, error) {
 	// It connects the OpenTelemetry Collector through local gRPC connection.
 	// You may replace `localhost:4317` with your endpoint.
@@ -102,7 +92,6 @@ func newTraceProvider(ctx context.Context, res *resource.Resource) (*trace.Trace
 
 	traceProvider := trace.NewTracerProvider(
 		trace.WithBatcher(traceExporter,
-			// Default is 5s. Set to 1s for demonstrative purposes.
 			trace.WithBatchTimeout(time.Second),
 		),
 		trace.WithResource(res),
@@ -118,7 +107,6 @@ func newMeterProvider(res *resource.Resource) (*metric.MeterProvider, error) {
 
 	meterProvider := metric.NewMeterProvider(
 		metric.WithReader(metric.NewPeriodicReader(metricExporter,
-			// Default is 1m. Set to 3s for demonstrative purposes.
 			metric.WithInterval(30*time.Second))),
 		metric.WithResource(res),
 	)
